fix(env:vscode): remove workspace file in favor of `settings.json`
Workspaces can be configured by the user according to their preferences.

`yarn sdks` automatically adds/updates the pertinent settings in
`settings.json`, not the `*.code-workspace` file. If the latter file is
present, some settings in `settings.json` will be ignored --
particularly the `typescript.*` and `files.associations` settings which
we care about.